Ambient music ticker AoD: make it compatibile with double tap sensor
some devices like the Pixel2xl have a specific double tap sensor for AoD
that handles the touches in this screen state and nothing is dispatched
to StatusBarWindowView, and the code is linked to the double tap to wake
option available in Settings - System - Gestures.

We can use the related onDoubleTap callback that Google already added
on purpose to manage double taps for their Now Playing feature and trigger
the skip track event when double tap to wake is enabled.

Change-Id: Iacb715b23b638f024232385b0e63048f4b841947
Signed-off-by: phenomenall <venkysaihot@gmail.com>