Remove NotificationVisibility storage pool
Based on recent measurements, storage pools of small objects are
generally not useful.

We ran into this because growing ArrayDeques are potentially very
unfriendly towards a generational GC, and racing recycle calls
can cause it it grow beyond MAX_POOL_SIZE. This isn't a big deal,
but there's no reason we should even be thinking about it.

Test: Build and boot AOSP

Change-Id: Icbda95e472e206c54f141129cc36ddc6d163b095
Signed-off-by: phenomenall <venkysaihot@gmail.com>